Add breeze run command
This adds a breeze run and breeze sh command. Breeze run
is effectively the same as kubectl run, with automatic pod
deletion after the run.

This works by first building the docker image, then starting a pod
with restart policy Never, and then attaching to the pod with stdin
and stdout attached.